requirements.txt
opencv-python == 4.9.0.80
opencv-contrib-python
numpy
qtpy
pyqt5
pyvistaqt
colour
qtawesome
cached-property
scipy
pandas
py_structs < 1.0
numba
matplotlib
simple-parsing
natsort
palettable
numpy-quaternion
omegaconf
packaging
tqdm